U.S. Presidential Polls and Equity Markets
Markets have been on an uptick and continue to keep momentum as we draw near to the US Presidential elections. Polls, like markets, are sentiment-based and we expect to see investor sentiments reflected in polls and vice versa. 
As such, for this project, we wish to verify if the US Presidential polls have explanatory value for the S&P stock market performance. In particular, we started off with the hypothesis that the incumbent's poll rates will be more likely to correlate with stock markets, i.e. Donald Trump's likelihood to win/lose may be more linked to stock markets.